pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spine"
version = "0.1.0"
description = "SPINE — local-first, repo-native mission governor CLI"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "HASHI.AI", email = "dev@hashi.ai" },
]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "rich>=13.0",
  "mcp>=1.0.0",
]

[project.scripts]
spine = "spine.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/spine"]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "ruff>=0.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]
